moved floating buttons
